% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_all.R
\name{cv_all}
\alias{cv_all}
\title{calculate the CV optimal beta}
\usage{
cv_all(X, y, D, option)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{D}{the linear transform}

\item{option}{options for screening}
}
\value{
beta_hat: CV optimal beta

stat_cv: various intermedia statistics
}
\description{
cv_all calculate the CV optimal beta
in the problem 1/n |y - X beta|^2 + 1/nu |D beta - gamma|^2 + lambda |gamma|_1.
}
