% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_screen.R
\name{cv_screen}
\alias{cv_screen}
\title{calculate the CV optimal beta and estimated support set}
\usage{
cv_screen(X, y, D, option)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{D}{the linear transform}

\item{option}{options for screening}
}
\value{
beta_hat: CV optimal beta

stat_cv: various intermedia statistics, including the estimated support sets
}
\description{
cv_all calculate the CV optimal beta and estimated
support set in the problem 1/n |y - X beta|^2 + 1/nu |D beta - gamma|^2 + lambda |gamma|_1.
}
