% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{normc}
\alias{normc}
\title{default normalization function for matrix}
\usage{
normc(X)
}
\arguments{
\item{X}{the input martix}
}
\value{
Y the output matrix
}
\description{
normalize columns of a matrix.
}
\examples{
library(mvtnorm)
n = 350
p = 100
Sigma = matrix(0, p, p)
X <- rmvnorm(n,matrix(0, p, 1), Sigma)
X <- normc(X)
}
