% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu_eval}
\alias{simu_eval}
\title{simu_eval}
\usage{
simu_eval(gamma_true, result)
}
\arguments{
\item{gamma_true}{true signal of gamma}

\item{result}{the estimated support set of gamma}
}
\value{
fdr: false discovery rate of the estimated support set

power: power of the estimated support set
}
\description{
calculate the FDR and Power when gamma^* is available, i.e. in simulations.
}
