% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu_unit}
\alias{simu_unit}
\title{default unit for simulations}
\usage{
simu_unit(n, p, D, A, c, k, option)
}
\arguments{
\item{n}{the sample size}

\item{p}{the dimension of variables}

\item{D}{the linear transform}

\item{A}{SNR}

\item{c}{feature correlation}

\item{k}{number of nonnulls in beta}

\item{option}{option for split knockoffs}
}
\value{
simu_data: a structure contains the following elements

simu_data$fdr_split: a vector recording fdr of split knockoffs w.r.t.nu

simu_data$power_split: a vector recording power of split knockoffs w.r.t.nu
}
\description{
the simulation unit for simulation experiments.
}
