% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu_unit_cv}
\alias{simu_unit_cv}
\title{simu_unit_cv}
\usage{
simu_unit_cv (n, p, D_s, A, c, k, option)
}
\arguments{
\item{n}{the sample size}

\item{p}{the dimension of variables}

\item{D_s}{the set of linear transform}

\item{A}{SNR}

\item{c}{feature correlation}

\item{k}{number of nonnulls in beta}

\item{option}{option for split knockoffs}
}
\value{
simu_data: a structure contains the following elements

simu_data$fdr: fdr of cv optimal nu in split knockoffs

simu_data$power: power of cv optimal nu in split knockoffs

simu_data$fdr_knock: fdr of knockoffs

simu_data$cv_list: length(D_s) * length(nu_s) matrix with cv loss for
each nu for each D

simu_data$chosen_nu: cv selected nu for each D
}
\description{
the simulation unit for simulation experiments with cross validation.
}
