% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{sk.filter}
\alias{sk.filter}
\title{split Knockoff filter for structural sparsity problem}
\usage{
sk.filter(X, D, y, option)
}
\arguments{
\item{X}{the design matrix}

\item{D}{the response vector}

\item{y}{the linear transformation}

\item{option}{options for creating the Split Knockoff statistics.
option$q: the desired FDR control target.
option$beta: choices on beta(lambda), can be: 'path', beta(lambda)
is taken from a regularization path; 'cv_beta', beta(lambda) is taken
as the cross validation optimal estimator hat beta; or 'cv_all',
beta(lambda) as well as nu are taken from the cross validation
optimal estimators hat beta and hat nu.The default setting is
'cv_all'.
option$lambda_cv: a set of lambda appointed for cross validation in
estimating hat beta, default 10.^seq(0, -8, by = -0.4).
option$nu_cv: a set of nu appointed for cross validation in
estimating hat beta and hat nu, default 10.^seq(0, 2, by = 0.4).
option$nu: a set of nu used in option.beta = 'path' or 'cv_beta' for
Split Knockoffs, default 10.^seq(0, 2, by = 0.2).
option$lambda: a set of lambda appointed for Split LASSO path
calculation, default 10.^seq(0, -6, by = -0.01).
option$normalize: whether to normalize the data, default true.
option$W: the W statistics used for Split Knockoffs, can be 's', 'st',
'bc', 'bct', default 'st'.}
}
\value{
various intermedia statistics
}
\description{
split Knockoff filter for structural sparsity problem
}
