% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{sk.select}
\alias{sk.select}
\title{split knockoff selector given W statistics}
\usage{
sk.select(W, q, option)
}
\arguments{
\item{W}{statistics W_j for testing null hypothesis}

\item{q}{target FDR}

\item{option}{option$method can be 'knockoff' or 'knockoff+'}
}
\value{
S array of selected variable indices
}
\description{
split knockoff selector given W statistics
}
