% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_fixed.R
\name{sk.W_fixed}
\alias{sk.W_fixed}
\title{W statistics generator based on a fixed beta(lambda) = hat beta}
\usage{
sk.W_fixed(X, D, y, nu, option)
}
\arguments{
\item{X}{the design matrix}

\item{D}{the linear transform}

\item{y}{the response vector}

\item{nu}{the parameter for variable splitting}

\item{option}{options for creating the Knockoff statistics
option$lambda: the choice of lambda for the path
option$beta_hat: the choice of beta(lambda) = hat beta}
}
\value{
the split knockoff statistics W and various intermedia statistics
}
\description{
generates the split knockoff statistics W based on a fixed beta(lambda) = hat beta in the
intercepetion assignment step.
}
