% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_path.R
\name{sk.W_path}
\alias{sk.W_path}
\title{W statistics generator based on the beta(lambda) from a split LASSO path}
\usage{
sk.W_path(X, D, y, nu, option)
}
\arguments{
\item{X}{the design matrix}

\item{D}{the linear transform}

\item{y}{the response vector}

\item{nu}{the parameter for variable splitting}

\item{option}{options for creating the Knockoff statistics
option$lambda: the choice of lambda for the path}
}
\value{
the split knockoff statistics W and various intermedia statistics
}
\description{
generates the split knockoff statistics W based on the beta(lambda) from
a split LASSO path in the intercepetion assignment step.
}
