% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{split_knockoffs.statistics.magnitude.cv_mag}
\alias{split_knockoffs.statistics.magnitude.cv_mag}
\title{split_knockoffs.statistics.magnitude.cv_mag}
\usage{
split_knockoffs.statistics.magnitude.cv_mag(X, y, D, nu, option)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{D}{the linear transform}

\item{nu}{the parameter for variable splitting}

\item{option}{options for calculating cv
option$lambda the choice of lambda for the path}
}
\value{
lambda: CV optimal lambda
}
\description{
calculate the CV optimal lambda in the problem
1/n |y - X beta|^2 + 1/nu |D beta - gamma|^2 + lambda |gamma|_1
for a fixed nu.
}
