% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{split_knockoffs.statistics.magnitude.W_mag}
\alias{split_knockoffs.statistics.magnitude.W_mag}
\title{split_knockoffs.statistics.magnitude.W_mag}
\usage{
split_knockoffs.statistics.magnitude.W_mag(X, D, y, nu, option)
}
\arguments{
\item{X}{the design matrix}

\item{D}{the linear transform}

\item{y}{the response vector}

\item{nu}{the parameter for variable splitting}

\item{option}{options for creating the Knockoff statistics
option$eta the choice of eta for creating the knockoff copy
option$lambda the choice of lambda for the path}
}
\value{
W: the knockoff statistics

Z: feature significance

t_Z: knockoff significance
}
\description{
generate the knockoff statistics W, using the method of magnitude.
lambda here is chosen by cross validation.
}
