% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{splitknockoff.decompose}
\alias{splitknockoff.decompose}
\title{splitknockoff.decompose}
\usage{
splitknockoff.decompose(X, randomize)
}
\arguments{
\item{X}{the input matrix}

\item{randomize}{whether to randomize}
}
\value{
U

S = S

V = V

U_perp =U_perp
}
\description{
splitknockoff.decompose
}
\examples{
library(mvtnorm)
n = 350
p = 100
Sigma = matrix(0, p, p)
X <- rmvnorm(n,matrix(0, p, 1), Sigma)
decompose.result <- splitknockoff.decompose(X)
U_perp <- decompose.result$U_perp
}
