% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{splitknockoff.select}
\alias{splitknockoff.select}
\title{splitknockoff.select}
\usage{
splitknockoff.select(W, q)
}
\arguments{
\item{W}{statistics W_j for testing null hypothesis}

\item{q}{target FDR}
}
\value{
S array of selected variable indices
}
\description{
splitknockoff.select
}
