% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{compute the threshold for variable selection}
\usage{
threshold(W, q, method = "knockoff+")
}
\arguments{
\item{W}{statistics W_j for testing null hypothesis beta_j = 0}

\item{q}{target FDR}

\item{method}{option$method can be 'knockoff' or 'knockoff+'}
}
\value{
T: threshold for variable selection
}
\description{
compute the threshold for variable selection
}
