% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_W.R
\name{W_sign}
\alias{W_sign}
\title{W statistics generator for directional FDR control}
\usage{
W_sign(X, D, y, nu, option)
}
\arguments{
\item{X}{the design matrix}

\item{D}{the linear transform}

\item{y}{the response vector}

\item{nu}{the parameter for variable splitting}

\item{option}{options for creating the Knockoff statistics
option$eta specify the choice of eta for creating the knockoff copy; 
option$lambda specify the choice of lambda for the split LASSO path}
}
\value{
W: the knockoff statistics

Z: feature significance

r: the sign estimator

t_Z: knockoff significance
}
\description{
generate the split knockoff statistics W for a split LASSO 
path, take the directional effect into account
}
