% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{hittingpoint}
\alias{hittingpoint}
\title{hitting point calculator on a given path}
\usage{
hittingpoint(coef, lambda_vec)
}
\arguments{
\item{coef}{the path for one variable}

\item{lambda_vec}{respective value of lambda in the path}
}
\value{
Z: the hitting time

r: the sign of respective variable at the hitting time
}
\description{
calculate the hitting time and the sign of
respective variable in a path.
}
