% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu_eval}
\alias{simu_eval}
\title{functions for evaluating simulations}
\usage{
simu_eval(gamma_true, result, r)
}
\arguments{
\item{gamma_true}{true signal of gamma}

\item{result}{the estimated support set of gamma}

\item{r}{the estimated directional effect}
}
\value{
fdr: false discovery rate of the estimated support set

power: power of the estimated support set
}
\description{
calculate the FDR and Power for simulations.
}
